Diabetes Prediction Using Machine Learning
In this project we are going to discuss how to use Machine Learning to predict Diabetes. The prediction will predict whether a patient has a diabetes or not. We have used ‘K-Nearest Neighbors’ (KNN) machine learning algorithm for this work. We also measure the Accuracy of K-NN the algorithm and also compare it with another algorithm that is Decision Tree algorithm. Algorithms are works with python Machine learning modules like pandas, sklearn, matplotlib, Seaborn, numpy etc.